Add dependency manager
