+++ golem-rib-repl/Cargo.toml
@@ -15,8 +15,13 @@ path = "src/main.rs"
 [dependencies]
 golem-rib = { path = "../golem-rib", version = "=0.0.0" }
 golem-worker-service-base = { path = "../golem-worker-service-base", version = "=0.0.0", optional = true }
+golem-wasm-ast = { path = "../wasm-ast", version = "=0.0.0", default-features = false, features = [
+    "analysis",
+    "bincode",
+] }
 
 rustyline = "15"
 termion = "1.5.6"
 tokio = { version = "1", features = ["full"] }
 dirs = "4.0.0"
+async-trait = "0.1.87"
